requirements.txt
web3==6.15.1
solana==0.34.3
solders==0.21.0
base58==2.1.1
pyyaml==6.0
aiohttp==3.9.1
requests==2.31.0
colorama==0.4.6
python-dotenv==1.0.0
flask==3.0.0
flask-socketio==5.3.4